User auth routes, closes #52
